backend/requirements.txt
alembic==1.16.2
altair==5.5.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
bokeh==3.7.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.0
contourpy==1.3.2
cryptography==44.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-DebugToolbar==0.16.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.2
fonttools==4.57.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwt==1.3.1
kiwisolver==1.4.8
limits==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.37.1
networkx==3.5
numpy==2.2.4
openai==1.76.0
ordered-set==4.1.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
plotly==6.0.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xyzservices==2025.4.0
